pyproject.toml
[project]
name = "get-ontology-findings"
version = "0.1.0"
description = "Get imaging findings and their codes from ontologies."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.21.0",
    "instructor>=1.7.8",
    "jupyter>=1.1.1",
    "openai>=1.68.2",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.2",
]
